+++ pkg/analysis_server/test/services/correction/organize_directives_test.dart
@@ -582,7 +582,7 @@ import 'package:b/a.dart';''');
 
   Future<void> _computeUnitAndErrors(String code) async {
     addTestSource(code);
-    var result = await session.getResolvedUnit2(testFile) as ResolvedUnitResult;
+    var result = await session.getResolvedUnit(testFile) as ResolvedUnitResult;
     testUnit = result.unit!;
     testErrors = result.errors;
   }