+++ pkg/analyzer/test/src/dart/analysis/session_test.dart
@@ -33,7 +33,7 @@ class AnalysisSessionImpl_BazelWorkspaceTest
 
     var path = convertPath('$workspaceRootPath/$relPath');
     var session = contextFor(path).currentSession;
-    var result = await session.getErrors2(path);
+    var result = await session.getErrors(path);
     expect(result, isA<NotPathOfUriResult>());
   }
 
@@ -57,7 +57,7 @@ class AnalysisSessionImpl_BazelWorkspaceTest
 
     var path = convertPath('$workspaceRootPath/$relPath');
     var session = contextFor(path).currentSession;
-    var result = session.getParsedLibrary2(path);
+    var result = session.getParsedLibrary(path);
     expect(result, isA<NotPathOfUriResult>());
   }
 
@@ -67,7 +67,7 @@ class AnalysisSessionImpl_BazelWorkspaceTest
 
     var path = convertPath('$workspaceRootPath/$relPath');
     var session = contextFor(path).currentSession;
-    var result = await session.getResolvedLibrary2(path);
+    var result = await session.getResolvedLibrary(path);
     expect(result, isA<NotPathOfUriResult>());
   }
 
@@ -77,7 +77,7 @@ class AnalysisSessionImpl_BazelWorkspaceTest
 
     var path = convertPath('$workspaceRootPath/$relPath');
     var session = contextFor(path).currentSession;
-    var result = await session.getResolvedUnit2(path);
+    var result = await session.getResolvedUnit(path);
     expect(result, isA<NotPathOfUriResult>());
   }
 
@@ -88,8 +88,7 @@ class AnalysisSessionImpl_BazelWorkspaceTest
     );
 
     var session = contextFor(file.path).currentSession;
-    var result =
-        await session.getResolvedUnit2(file.path) as ResolvedUnitResult;
+    var result = await session.getResolvedUnit(file.path) as ResolvedUnitResult;
     expect(result.state, ResultState.VALID);
     expect(result.path, file.path);
     expect(result.errors, isEmpty);
@@ -103,7 +102,7 @@ class AnalysisSessionImpl_BazelWorkspaceTest
     );
 
     var session = contextFor(file.path).currentSession;
-    var result = await session.getUnitElement2('not_absolute.dart');
+    var result = await session.getUnitElement('not_absolute.dart');
     expect(result, isA<InvalidPathResult>());
   }
 
@@ -113,7 +112,7 @@ class AnalysisSessionImpl_BazelWorkspaceTest
 
     var path = convertPath('$workspaceRootPath/$relPath');
     var session = contextFor(path).currentSession;
-    var result = await session.getUnitElement2(path);
+    var result = await session.getUnitElement(path);
     expect(result, isA<NotPathOfUriResult>());
   }
 
@@ -175,12 +174,12 @@ test:lib/
   }
 
   test_getErrors2_invalidPath_notAbsolute() async {
-    var errorsResult = await session.getErrors2('not_absolute.dart');
+    var errorsResult = await session.getErrors('not_absolute.dart');
     expect(errorsResult, isA<InvalidPathResult>());
   }
 
   test_getFile2_invalidPath_notAbsolute() async {
-    var errorsResult = session.getFile2('not_absolute.dart');
+    var errorsResult = session.getFile('not_absolute.dart');
     expect(errorsResult, isA<InvalidPathResult>());
   }
 
@@ -216,7 +215,7 @@ class B {}
   }
 
   test_getLibraryByUri2_unresolvedUri() async {
-    var result = await session.getLibraryByUri2('package:foo/foo.dart');
+    var result = await session.getLibraryByUri('package:foo/foo.dart');
     expect(result, isA<CannotResolveUriResult>());
   }
 
@@ -242,13 +241,13 @@ class B {}
   }
 
   test_getParsedLibrary2_invalidPath_notAbsolute() async {
-    var result = session.getParsedLibrary2('not_absolute.dart');
+    var result = session.getParsedLibrary('not_absolute.dart');
     expect(result, isA<InvalidPathResult>());
   }
 
   test_getParsedLibrary2_notLibrary() async {
     newFile(testPath, content: 'part of "a.dart";');
-    expect(session.getParsedLibrary2(testPath), isA<NotLibraryButPartResult>());
+    expect(session.getParsedLibrary(testPath), isA<NotLibraryButPartResult>());
   }
 
   test_getParsedLibrary2_parts() async {
@@ -332,7 +331,7 @@ class B {}
     newFile(testPath, content: '');
 
     var resolvedUnit =
-        await session.getResolvedUnit2(testPath) as ResolvedUnitResult;
+        await session.getResolvedUnit(testPath) as ResolvedUnitResult;
     var typeProvider = resolvedUnit.typeProvider;
     var intClass = typeProvider.intType.element;
 
@@ -417,7 +416,7 @@ part 'c.dart';
     var aaaSession =
         contextCollection.contextFor(aaaContextPath).currentSession;
 
-    var result = aaaSession.getParsedLibraryByElement2(element);
+    var result = aaaSession.getParsedLibraryByElement(element);
     expect(result, isA<NotElementOfThisSessionResult>());
   }
 
@@ -435,7 +434,7 @@ class B {}
   }
 
   test_getParsedUnit2_invalidPath_notAbsolute() async {
-    var result = session.getParsedUnit2('not_absolute.dart');
+    var result = session.getParsedUnit('not_absolute.dart');
     expect(result, isA<InvalidPathResult>());
   }
 
@@ -506,14 +505,14 @@ class B2 extends X {}
   }
 
   test_getResolvedLibrary2_invalidPath_notAbsolute() async {
-    var result = await session.getResolvedLibrary2('not_absolute.dart');
+    var result = await session.getResolvedLibrary('not_absolute.dart');
     expect(result, isA<InvalidPathResult>());
   }
 
   test_getResolvedLibrary2_notLibrary() async {
     newFile(testPath, content: 'part of "a.dart";');
 
-    var result = await session.getResolvedLibrary2(testPath);
+    var result = await session.getResolvedLibrary(testPath);
     expect(result, isA<NotLibraryButPartResult>());
   }
 
@@ -603,7 +602,7 @@ part 'c.dart';
     var aaaSession =
         contextCollection.contextFor(aaaContextPath).currentSession;
 
-    var result = await aaaSession.getResolvedLibraryByElement2(element);
+    var result = await aaaSession.getResolvedLibraryByElement(element);
     expect(result, isA<NotElementOfThisSessionResult>());
   }
 
@@ -614,7 +613,7 @@ class B {}
 ''');
 
     var unitResult =
-        await session.getResolvedUnit2(testPath) as ResolvedUnitResult;
+        await session.getResolvedUnit(testPath) as ResolvedUnitResult;
     expect(unitResult.session, session);
     expect(unitResult.path, testPath);
     expect(unitResult.uri, Uri.parse('package:test/test.dart'));
@@ -643,39 +642,39 @@ class B {}
 
 extension on AnalysisSession {
   Future<UnitElementResult> getUnitElementValid(String path) async {
-    return await getUnitElement2(path) as UnitElementResult;
+    return await getUnitElement(path) as UnitElementResult;
   }
 
   ParsedLibraryResult getParsedLibraryValid(String path) {
-    return getParsedLibrary2(path) as ParsedLibraryResult;
+    return getParsedLibrary(path) as ParsedLibraryResult;
   }
 
   FileResult getFileValid(String path) {
-    return getFile2(path) as FileResult;
+    return getFile(path) as FileResult;
   }
 
   ParsedUnitResult getParsedUnitValid(String path) {
-    return getParsedUnit2(path) as ParsedUnitResult;
+    return getParsedUnit(path) as ParsedUnitResult;
   }
 
   Future<ResolvedLibraryResult> getResolvedLibraryValid(String path) async {
-    return await getResolvedLibrary2(path) as ResolvedLibraryResult;
+    return await getResolvedLibrary(path) as ResolvedLibraryResult;
   }
 
   Future<LibraryElementResult> getLibraryByUriValid(String path) async {
-    return await getLibraryByUri2(path) as LibraryElementResult;
+    return await getLibraryByUri(path) as LibraryElementResult;
   }
 
   Future<ResolvedLibraryResult> getResolvedLibraryByElementValid(
       LibraryElement element) async {
-    return await getResolvedLibraryByElement2(element) as ResolvedLibraryResult;
+    return await getResolvedLibraryByElement(element) as ResolvedLibraryResult;
   }
 
   ParsedLibraryResult getParsedLibraryByElementValid(LibraryElement element) {
-    return getParsedLibraryByElement2(element) as ParsedLibraryResult;
+    return getParsedLibraryByElement(element) as ParsedLibraryResult;
   }
 
   Future<ErrorsResult> getErrorsValid(String path) async {
-    return await getErrors2(path) as ErrorsResult;
+    return await getErrors(path) as ErrorsResult;
   }
 }