+++ pkg/nnbd_migration/lib/src/front_end/info_builder.dart
@@ -96,7 +96,7 @@ class InfoBuilder {
     for (var filePath in sources) {
       progressBar.tick();
       var session = driverProvider!.getAnalysisSession(filePath);
-      var result = await session.getResolvedLibrary2(filePath!);
+      var result = await session.getResolvedLibrary(filePath!);
       if (result is ResolvedLibraryResult) {
         for (var unitResult in result.units) {
           var sourceInfo =