+++ pkg/nnbd_migration/test/api_test.dart
@@ -65,7 +65,7 @@ abstract class _ProvisionalApiTestBase extends AbstractContextTest {
         removeViaComments: removeViaComments,
         warnOnWeakCode: warnOnWeakCode);
     for (var path in input.keys) {
-      var resolvedLibrary = await session.getResolvedLibrary2(path);
+      var resolvedLibrary = await session.getResolvedLibrary(path);
       if (resolvedLibrary is ResolvedLibraryResult) {
         for (var unit in resolvedLibrary.units) {
           var errors =
@@ -80,7 +80,7 @@ abstract class _ProvisionalApiTestBase extends AbstractContextTest {
     expect(migration.unmigratedDependencies, isEmpty);
     _betweenStages();
     for (var path in input.keys) {
-      var resolvedLibrary = await session.getResolvedLibrary2(path);
+      var resolvedLibrary = await session.getResolvedLibrary(path);
       if (resolvedLibrary is ResolvedLibraryResult) {
         for (var unit in resolvedLibrary.units) {
           migration.processInput(unit);
@@ -89,7 +89,7 @@ abstract class _ProvisionalApiTestBase extends AbstractContextTest {
     }
     _betweenStages();
     for (var path in input.keys) {
-      var resolvedLibrary = await session.getResolvedLibrary2(path);
+      var resolvedLibrary = await session.getResolvedLibrary(path);
       if (resolvedLibrary is ResolvedLibraryResult) {
         for (var unit in resolvedLibrary.units) {
           migration.finalizeInput(unit);