+++ pkg/analyzer/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Changed `ParsedLibraryResult.units` to be non-nullable.
 * Changed `ResolvedLibraryResult.element` to be non-nullable.
 * Changed `ResolvedLibraryResult.units` to be non-nullable.
+* Deprecated and renamed `AnalysisSession.getXyz2()` into `getXyz()`.
 
 ## 2.0.0
 * Removed deprecated `Scope.lookup2()`.